ppc64: add memreserve sections
to correctly perform a kexec reboot, we require to mark some of reserve
mem sections in the dtb blob. we compute the required information
similer to existing function `create_flatten_tree` and recursing over
/proc/device-tree directory.

adding /memreserve/ information inside dtb,
requires us to resize and reopen the into a new buffer.

Signed-off-by: Shivang Upadhyay <shivangu@linux.ibm.com>